Make protected calls to isDown() functions.
Since Love 0.11 love.keyboard.isDown() and love.mouse.isDown()
raise exceptions in case of key mismatch.

The current code relies on the old behaviour where no error was signaled
and the function simply returned false.

Protect those calls with pcall() to check if an exception is raised when
the aforementioned functions are called.

The change is backward compatible with older versions of Love.